+++ inbm/cloudadapter-agent/cloudadapter/cloud/client/inbs_cloud_client.py
@@ -115,8 +115,8 @@ class InbsCloudClient(CloudClient):
         @param time: timestamp for this telemetry publish
         @exception PublishError: If publish fails
         """
-
-        pass  # INBS is not yet ready to receive telemetry
+        
+        pass  # INBS is not yet ready to receive events
 
     def publish_node_update_response(self, key: str, value: str) -> None:
         """Publishes a response to the node update message to the dispatcher
@@ -171,9 +171,10 @@ class InbsCloudClient(CloudClient):
         
         request = inbs_sb_pb2.SendNodeUpdateRequest(
             request_id=str(uuid.uuid4()),
-            job_update=job,            
+            job_update=job,
+            static_telemetry=None,         
         )
-        logger.debug(f"Sending node update to INBS: request={request}")
+        logger.debug(f"Sending node update job status to INBS: request={request}")
             
         try:
             response = self._grpc_channel.SendNodeUpdate(request, metadata=self._metadata)
@@ -195,11 +196,35 @@ class InbsCloudClient(CloudClient):
         """Publishes a device attribute to the cloud
 
         @param key: attribute's key
-        @param value: value to set for the attribute
+        @param value: attribute's value
         @exception PublishError: If publish fails
         """
 
-        pass  # INBS is not yet ready to receive attributes
+        logger.debug(f"Received telemetry: key={key}, value={value}")
+        
+        if self._grpc_channel is None:
+            raise PublishError("gRPC channel not set up before calling InbsCloudClient.publish_node_update")            
+
+        
+        static_telemetry=common_pb2.StaticTelemetry(
+            node_id=self._client_id,
+        ) 
+        
+        static_telemetry.key = key
+        static_telemetry.value = value       
+         
+        request = inbs_sb_pb2.SendNodeUpdateRequest(
+            request_id=str(uuid.uuid4()),
+            job_update=None,
+            static_telemetry=static_telemetry,            
+        )
+        logger.debug(f"Sending node update of static telemetry to INBS: request={request}")
+            
+        try:
+            response = self._grpc_channel.SendNodeUpdate(request, metadata=self._metadata)
+            logger.info(f"Received response from gRPC server: {response}")
+        except grpc.RpcError as e:
+            logger.error(f"Failed to send node update via gRPC: {e}")
 
     def bind_callback(self, name: str, callback: Callable) -> None:
         """Bind a callback to be triggered by a method called on the cloud