+++ inbm/Changelog.md
@@ -3,7 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## NEXT - YYYY-MM-DD
+## 4.2.8 - 2024-11-29
+### Changed
+ - (NEXMANAGE-513) Record the granular log in download-only mode
+ - (NEXMANAGE-1022) Adds Power Capabilities to Static Telemetry data.  Sends Static Telemetry data to INBS via the INBS Cloud Client in CloudAdapter.
+   
+### Added
+- (NEXMANAGE-826) Update granular log to support architecture independent package
+
+## 4.2.7 - 2024-11-8
 ### Changed
  - (NEXMANAGE-949) Check the IMAGE_BUILD_DATE instead of VERSION for TiberOS SOTA verification
 
@@ -12,6 +20,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
  - (NEXMANAGE-733) Fixed typo in package_list during conversion from INBS to cloudadapter
+ - (NEXMANAGE-827) Fixed error logged incorrectly when reading dispatcher_state file
 
 ## 4.2.6.2 - 2024-10-25
 ### Added
@@ -37,6 +46,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
  - (NEXMANAGE-874) Remove UT rollback command in dispatcher agent
  - (NEXMANAGE-853) Remove ORAS and update method
+ - (NEXMANAGE-683) Remove service_type from Operation in common.proto
 
 ## 4.2.6 - 2024-10-04
 ### Added
@@ -64,6 +74,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - (NEXMANAGE-744) Remove `psutil` in favor if `shutil.disk_usage` to save space
  - (NEXMANAGE-744) Don't pull in `dmidecode` in `inbm-lib` -- pull in only in `telemetry`
  - (NEXMANAGE-744) Removed all references to `future` library as we do not use Python 2
+
 
 ## 4.2.5 - 2024-09-04
 ### Fixed